Move test resource functions away from the main lib
Test code should be separate to the main library code.
This moves code relating to identifying paths at which test files are located.
