+++ iati/tests/test_resources.py
@@ -66,7 +66,7 @@ class TestResourceFolders(object):
             Deal with multiple versions.
 
         """
-        paths = iati.resources.get_test_data_paths_in_folder('ssot-activity-xml-fail')
+        paths = iati.tests.resources.get_test_data_paths_in_folder('ssot-activity-xml-fail')
 
         assert len(paths) == 237
 
@@ -203,7 +203,7 @@ class TestResourceLoading(object):
 
     def test_load_as_bytes(self):
         """Test that resources.load_as_bytes returns a bytes object with the expected content."""
-        path_test_data = iati.resources.get_test_data_path('invalid')
+        path_test_data = iati.tests.resources.get_test_data_path('invalid')
 
         result = iati.resources.load_as_bytes(path_test_data)
 
@@ -212,7 +212,7 @@ class TestResourceLoading(object):
 
     def test_load_as_dataset(self):
         """Test that resources.load_as_dataset returns a Dataset object with the expected content."""
-        path_test_data = iati.resources.get_test_data_path('valid')
+        path_test_data = iati.tests.resources.get_test_data_path('valid')
 
         result = iati.resources.load_as_dataset(path_test_data)
 
@@ -221,14 +221,14 @@ class TestResourceLoading(object):
 
     def test_load_as_dataset_invalid(self):
         """Test that resources.load_as_dataset raises an error when the provided path does not lead to a file containing valid XML."""
-        path_test_data = iati.resources.get_test_data_path('invalid')
+        path_test_data = iati.tests.resources.get_test_data_path('invalid')
 
         with pytest.raises(ValueError):
             _ = iati.resources.load_as_dataset(path_test_data)
 
     def test_load_as_string(self):
         """Test that `resources.load_as_string()` returns a string (python3) or unicode (python2) object with the expected content."""
-        path_test_data = iati.resources.get_test_data_path('invalid')
+        path_test_data = iati.tests.resources.get_test_data_path('invalid')
 
         result = iati.resources.load_as_string(path_test_data)
 
@@ -238,7 +238,7 @@ class TestResourceLoading(object):
     @pytest.mark.parametrize("load_method", [iati.resources.load_as_bytes, iati.resources.load_as_dataset, iati.resources.load_as_string])
     def test_load_as_x_non_existing_file(self, load_method):
         """Test that `resources.load_as_bytes()` returns a bytes object with the expected content."""
-        path_test_data = iati.resources.get_test_data_path('this-file-does-not-exist')
+        path_test_data = iati.tests.resources.get_test_data_path('this-file-does-not-exist')
 
         # python 2/3 compatibility - FileNotFoundError introduced at Python 3
         try:
@@ -254,7 +254,7 @@ class TestResourceLoading(object):
     ])
     def test_load_as_string_restricted_charset(self, file_to_load):
         """Test that Datasets can be loaded from files encoded with a limited charset."""
-        path = iati.resources.get_test_data_path(file_to_load)
+        path = iati.tests.resources.get_test_data_path(file_to_load)
 
         data_str = iati.resources.load_as_string(path)
         dataset = iati.Dataset(data_str)
@@ -273,7 +273,7 @@ class TestResourceLoading(object):
     ])
     def test_load_as_string_unicode(self, file_to_load):
         """Test that Datasets can be loaded from files encoded with various unicode encodings."""
-        path = iati.resources.get_test_data_path(file_to_load)
+        path = iati.tests.resources.get_test_data_path(file_to_load)
 
         data_str = iati.resources.load_as_string(path)
         dataset = iati.Dataset(data_str)
@@ -294,7 +294,7 @@ class TestResourceLoading(object):
         The file with an undetectable encoding is a UTF-16LE file without a BOM.
 
         """
-        path = iati.resources.get_test_data_path(file_to_load)
+        path = iati.tests.resources.get_test_data_path(file_to_load)
 
         with pytest.raises(ValueError):
             _ = iati.resources.load_as_string(path)
